mobitrash/payments/views.py
from django.shortcuts import render
from rest_framework import viewsets, filters
from .models import Payment
from .serializers import PaymentSerializer

class PaymentViewSet(viewsets.ModelViewSet):
    queryset = Payment.objects.all()
    serializer_class = PaymentSerializer
    filter_backends = (filters.DjangoFilterBackend,)
    filter_fields = ('user', 'status')
    search_fields = ('user__username', 'user__email', 'status')
    
    def get_queryset(self):
        return Payment.objects.filter(user=self.request.user)
    
    def perform_create(self, serializer):
        serializer.save(user=self.request.user)
        
