mobitrash/requirements.txt
beautifulsoup4==4.12.3
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
django-cors-headers==3.13.0 
django-filter==24.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3